feat: add update/edge-function command
