+++ messages/edge_function/messages.go
@@ -48,14 +48,15 @@ var (
 	ListHelpFlag         = "Displays more information about the list edge-function command"
 
 	//update cmd
-	UpdateUsage            = "update --function-id <function_id> [flags]"
-	UpdateShortDescription = "Modifies an Edge Function"
-	UpdateLongDescription  = "Modifies an Edge Function based on its ID to update its name, activity status, code path, and other attributes"
-	UpdateFlagName         = "The Edge Function's name"
-	UpdateFlagCode         = "Path and name to the file containing the Edge Function's code"
-	UpdateFlagActive       = "Whether the Edge Function should be active or not"
-	UpdateFlagArgs         = "Path and name of the JSON file containing the Edge Function's arguments"
-	UpdateFlagIn           = "Given path and JSON file to automatically update the Edge Function attributes; you can use - for reading from stdin"
-	UpdateOutputSuccess    = "Updated Edge Function with ID %d\n"
-	UpdateHelpFlag         = "Displays more information about the update edge-function command"
+	UpdateUsage             = "update --function-id <function_id> [flags]"
+	UpdateShortDescription  = "Modifies an Edge Function"
+	UpdateLongDescription   = "Modifies an Edge Function based on its ID to update its name, activity status, code path, and other attributes"
+	UpdateFlagName          = "The Edge Function's name"
+	UpdateFlagCode          = "Path and name to the file containing the Edge Function's code"
+	UpdateFlagActive        = "Whether the Edge Function should be active or not"
+	UpdateFlagArgs          = "Path and name of the JSON file containing the Edge Function's arguments"
+	UpdateFlagIn            = "Given path and JSON file to automatically update the Edge Function attributes; you can use - for reading from stdin"
+	UpdateOutputSuccess     = "Updated Edge Function with ID %d\n"
+	UpdateHelpFlag          = "Displays more information about the update edge-function command"
+	UpdateAskEdgeFunctionID = "What is the ID of the edge function you wish to update?"
 )
