+++ pkg/cmd/update/edge_function/fixtures/update.json
