+++ pkg/cmd/update/update.go
@@ -6,6 +6,7 @@ import (
 	cacheSetting "github.com/aziontech/azion-cli/pkg/cmd/update/cache_setting"
 	domain "github.com/aziontech/azion-cli/pkg/cmd/update/domain"
 	edgeApplication "github.com/aziontech/azion-cli/pkg/cmd/update/edge_application"
+	edgeFunction "github.com/aziontech/azion-cli/pkg/cmd/update/edge_function"
 	origin "github.com/aziontech/azion-cli/pkg/cmd/update/origin"
 	rulesEngine "github.com/aziontech/azion-cli/pkg/cmd/update/rules_engine"
 
@@ -30,6 +31,7 @@ func NewCmd(f *cmdutil.Factory) *cobra.Command {
 	cmd.AddCommand(rulesEngine.NewCmd(f))
 	cmd.AddCommand(domain.NewCmd(f))
 	cmd.AddCommand(origin.NewCmd(f))
+	cmd.AddCommand(edgeFunction.NewCmd(f))
 	cmd.AddCommand(cacheSetting.NewCmd(f))
 
 	cmd.Flags().BoolP("help", "h", false, msg.FlagHelp)
